chore: move main.js to main directory
