+++ app/main/windows.js
@@ -1,7 +1,7 @@
 import {BrowserWindow, Menu, dialog, ipcMain, webContents} from 'electron';
 
 import i18n from '../configs/i18next.config';
-import {openFilePath} from '../main';
+import {openFilePath} from './main';
 import settings from '../shared/settings';
 import {APPIUM_SESSION_EXTENSION} from './helpers';
 import {rebuildMenus} from './menus';
